requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
idna==3.2
requests==2.26.0
soupsieve==2.2.1
tqdm==4.62.2
urllib3==1.26.6
anikimiapi==0.1.4
termcolor==1.3
python-mpv==0.5.2
